requirements.txt
# Only ONE external package needed — Gemini SDK
# Gmail uses Python's built-in imaplib (no install required)

google-generativeai
